Move files to campaign-actions folder
